Added support for Docker config parameter.
This commit adds support for the Docker CLI --config flag, passed as a
URL parameter. We've put this in the DaemonConnectionFlags type for now,
but we haven't wired it up to Compose V1 support (since it's only
supported by Compose V2). We'll likely remove DaemonConnectionFlags when
we remove Compose V1 support, since we can just do a more direct
conversion between parameters and CLI flags in the Docker transport.

Signed-off-by: Jacob Howard <jacob@mutagen.io>