Added internal staging mode.
This commit adds a new "internal" staging mode where files are
synchronized directly into a subdirectory of the staging root. This
actually works better than expected because the synchronization loop
will retry on missing files, meaning that even if the synchronization
root doesn't initially exist, it will still be created by transition,
and thus staging will succeed on the next synchronization cycle.
However, this isn't super efficient, because an entire staging operation
will have to fail, so we may want to figure out a mechanism to push
failure notifications earlier, but that's fine because this staging
mechanism is not something we'll recommend in general due to its strict
requirements.

Signed-off-by: Jacob Howard <jacob@mutagen.io>