PoC: sync: add sudo support
NOTE This is a PoC just to start a discussion on wether such feature could be
     interesting. It doesn't have tests and might require refactoring.

Sometimes you need to sync files that the SSH user cannot modify, and the beta
endpoint doesn't have root SSH access.

If the SSH user has passwordless sudo, adding to mutagen.yml

    sync:
      defaults:
        permissions:
          sudo: true

will invoke the mutagen agent with sudo.

Signed-off-by: Marco Molteni <marco.molteni@mailbox.org>