Added flag to control executable lookup behavior for daemon start.
The existing start (and autostart) code assumes that the calling
executable is the Mutagen CLI, but for external tools (such as Mutagen
Compose) calling in via cmd packages, this may not be true, so this
commit adds a flag (and a package for future flags) that external tools
can set to control Mutagen's behavior in this and other regards.

A corresponding change will be required in Mutagen Compose.

Updates mutagen-io/mutagen-compose#5.

Signed-off-by: Jacob Howard <jacob@mutagen.io>