Simplified sidecar tag script.
For the time being, we'll avoid doing complicated multi-tag sidecar
releases, because Mutagen Compose won't make use of them and other
projects will likely want to maintain similar precision tags. As Mutagen
matures and the sidecar matures, we might want to rethink this, but for
the time being it makes sense to only support pinning to specific
version tags.

Signed-off-by: Jacob Howard <jacob@mutagen.io>