ci: add non-release sidecar build
When running CI for a release, we don't want to push the sidecar image
to Docker Hub until we're sure that individual platform builds and tests
have succeeded (because if they haven't, we'll likely need to redo the
release). However, when running CI for a non-release build, it's tedious
to wait for all of the platforms to finish before testing the sidecar
build. It would be nice if we could make the sidecar build conditionally
dependent on the platform builds, but unfortunately GitHub Actions
doesn't provide a mechanism for that. Thus, this commit splits the
sidecar job into two: one for regular CI builds and one for release CI
builds. In the latter case, the sidecar build is dependent on the
platform builds completing successfully, whereas in the former case it
does not. Every effort has been made to keep the two jobs in sync in
terms of configuration.

Signed-off-by: Jacob Howard <jacob@mutagen.io>