ci: simplify commit verification
The commitlint tool seemed a little too flakey, incorrectly parsing some
commit bodies for reasons unknown. Even outside of the issues we
experienced, it seems like there are enough open parsing issues there to
make that type of automated approach fraught with peril.

So, this commit simplifies commit message format verification to simply
checking the maximum line length. Sign-off and cryptographic checks are
still in place. Format checks will have to be performed by a human, but
that's probably for the best, as it also avoids the need to maintain a
complex and probably ever growing list of type tags.

Signed-off-by: Jacob Howard <jacob@mutagen.io>