sync: fix suboptimal code scoping
For some reason we were computing error results before determining
whether or not there was actually an error. This code probably just
resulted accidentally from multiple refactors, but it was doing extra
work on every Transition operation in cases where a maximum entry count
was set (which, fortunately, is not too common).

Signed-off-by: Jacob Howard <jacob@mutagen.io>