sync: fix missed poll event with fast change inversion
With poll-based watching, if a change performed by Transition was fully
inverted by some external mechanism between polling operations, no
polling event was generated. This is because the pre-Transition and
post-Transition scans looked identical, so the controller was never
asked to look at them again. This was most evident in the case where a
new file was created, propagated, and then removed in the window between
poll-based scans, but it could occur with more complex inversions as
well (e.g. replacing a deleted file, unmodifying file contents, etc.).

The remedy to this is to track whether or not transition successfully
performs any on-disk changes and (if using poll-based watching) signal a
polling event in that case.

Signed-off-by: Jacob Howard <jacob@mutagen.io>