sync: refactor recursive and non-recursive watching APIs
The existing recursive watching API was a little weird for two reasons:

First, it used a somewhat free-hanging Goroutine that was hard to track
and manage. This also required an additional polling step to know when
(and indeed if) a watch had been successfully established.

Second, it placed the burden of event coalescing on the caller. In
addition to adding a lot of complexity to the endpoint implementation,
this ended up duplicating coalescing on macOS (which already has a
built-in, OS-provided coalescing mechanism).

This commit solves both issues. It adds a new RecursiveWatcher interface
that better abstracts and regulates watching, in addition to performing
event coalescing internally.

The non-recursive API was a little bit better, but it also made sense to
abstract into a NonRecursiveWatcher interface (with event coalescing)
for better parity with the recursive API. Doing this also allowed us to
update watch_demo to work with non-recursive watching platforms.

Signed-off-by: Jacob Howard <jacob@mutagen.io>