ci: disable SHA-256 digests and GPG signing for non-tag builds
If we're not doing a release (i.e. not building a tag), then disable the
GPG signing key import and SHA-256 digests in the macOS build. The main
motivation behind this change is to make the macOS build work for pull
requests originating from third-party repositories (where the GPG
secrets aren't available). While this could probably be accomplished
without fully disabling the SHA-256 digest computation, it's cleaner to
just disable it completely to keep consistency with macOS notarization
and because we don't need SHA-256 digests for development builds.

Signed-off-by: Jacob Howard <jacob@mutagen.io>