sync: modify FSEvents watching to use additional coalescing
It turns out that FSEvents' built-in coalescing is somewhat limited.
Specifically, it will only coalesce up to 32 events into a single event
set. In order to keep consistency with coalescing on other platforms,
this commit adds an additional coalescing layer on top of that provided
by FSEvents. This will yield better coalescing without sacrificing
latency (since we use kFSEventStreamCreateFlagNoDefer).

This commit also bumps the coalescing window to 20 ms, which seems to
give significantly better coalescing in practice.

Signed-off-by: Jacob Howard <jacob@mutagen.io>