sync: add filtering mechanism to watchers
This commit adds a filtering mechanism to watchers that can be used to
offload temporary file ignoring from endpoints to watchers. This not
only removes complexity from the endpoint implementation, but also makes
the watcher's communication with the endpoint more efficient.

This change may also lay the groundwork for applying session ignores to
watching itself.

Signed-off-by: Jacob Howard <jacob@mutagen.io>