sync: adopt new watching and scanning APIs in synchronization
This commit adopts the new watching and scanning APIs in the
synchronization controller and endpoints.

This refactoring is a follow up to #319, which is what originally
motivated these additional changes. Most of the ideas of #319 are still
present, though certain techniques (like transition change detection)
have been refined and/or simplified. The most important change is that
scan acceleration is no longer always disabled in Transition for
recursive watching.

Signed-off-by: Jacob Howard <jacob@mutagen.io>