ci: switch back to unified sidecar build
In 504d325, we split the sidecar build into two different jobs: one for
release builds and one for non-release builds. This was largely
motivated by debugging the sidecar container build (since having to wait
for other steps to complete was tedious), but this should just be done
by using a temporary commit with the build dependencies removed. So this
commit reunifies the sidecar build into a single step, which is easier
to manage and significantly less error-prone.

If GitHub Actions' workflow syntax eventually supports conditional
dependencies, that would be the ideal mechanism to use.

Signed-off-by: Jacob Howard <jacob@mutagen.io>