scan_bench: add second full warm scan
Statistically speaking, the most common scan mode when Mutagen is
polling is going to be a full warm scan after another full warm scan.
This may have significantly different characteristics than a full warm
scan after a full cold scan because the stat cache may be better
populated. In any case, it makes sense the measure and profile both
cases for the purposes of optimization.

Signed-off-by: Jacob Howard <jacob@mutagen.io>