sync: make directory seeks lazy
Mutagen's scan operations never need to read a directory's contents
twice. However, for the purposes of API completeness and correctness,
the POSIX Directory implementation always seeks the directory descriptor
to the beginning before reading directory contents. This isn't necessary
for a freshly opened directory descriptor and it can be deferred to
subsequent reads if needed.

Signed-off-by: Jacob Howard <jacob@mutagen.io>