sync: use O_DIRECTORY to enforce directory open type check
We previously used the same fstat mechanism that we use to enforce file
type checks when performing an openat operation, but we can instead use
O_DIRECTORY to enforce this as part of the openat call, ideally saving
us a system call when opening a directory. Unfortunately, this doesn't
seem to make a performance difference in practice (probably because the
kernel is doing something analagous anyway), but it's a better design
and benefits from any OS-level optimization in the future.

There's no benefit to doing this for the package-level Open function
because it has to use fstat to type-switch anyway.

Signed-off-by: Jacob Howard <jacob@mutagen.io>