sync: replace pathJoin to amortize string concatenation
The pathJoin function typically ended up concatenating three strings:
the parent path, the separator, and the leaf. While it might seem like a
micro-optimization to amortize the cost of the parent-separator
concatenation, benchmarks show that it shaves off about 4% from scan
times. It would be nice if this abstraction could be kept, but the
speedup seems worth it. The other path utilities don't perform any
allocation and are already about as optimized as possible, so they'll be
left in place. The risk of breakage from this inlining/amortization is
extremely minimal given how simple it is and how baked into the
synchronization model (i.e. tested) it is at this point.

Signed-off-by: Jacob Howard <jacob@mutagen.io>