DO NOT MERGE: sync: perform fstatat in parallel
This commit adds parallel invocation of fstatat in scans. Unfortunately,
it doesn't currently yield the expected boost in performance. Perhaps
this is simply too low-level of a point to perform this parallelization,
as it relies on directories having somewhat large numbers of files.
Other tools that perform parallel stat operations typically do so on
large arrays of paths (though we can't do that since we're using fstatat
and have limitations on open parent descriptors).

There could also be some weird interaction with the Go runtime and the
way that syscalls are performed. In any case, it doesn't seem to help
and actually seems to hurt, so we'll store it now for posterity.

In any case, the parallelism package seems useful.

Signed-off-by: Jacob Howard <jacob@mutagen.io>