sync: relocate accelerated scan cache backfill
The cache backfilling that occurs in accelerated scanning makes more
sense to perform in the directory handler, where the baseline entry is
actually used. By doing this, we can avoid a map lookup for colliding
cache entries by restricting the entry walking to only the baseline
entry. This should make a very minor improvement to performance, the
size of which will depend on how much baseline content is being reused.

Signed-off-by: Jacob Howard <jacob@mutagen.io>