sync: reuse ancestor for remote snapshot baseline until remote non-empty
The first scan from a remote without any content is going to return a
somewhat useless set of baseline bytes, especially when it comes to
performing the first, large differential snapshot transfer after initial
synchronization. So, we'll avoid storing baseline bytes (and thus
continue to use the ancestor as a baseline) until the remote has
demonstrated that it has a non-zero amount of on-disk content.

This commit also adds logging to the remote endpoint client to verify
that snapshot delta transfers are working correctly.

Signed-off-by: Jacob Howard <jacob@mutagen.io>