sync: add shorthand optimization for remote staging paths
In the vast majority of staging operations, especially large initial
staging operations, the endpoint is going to require all of the
requested paths. In this scenario, there's not point in sending all of
them back over the wire, so this commit introduces a shorthand into the
remote synchronization endpoint protocol to sidestep this overhead.

In theory, we could also look at using path indices rather than paths
themselves, but that would affect the signature of the Stage method
itself (or require additional computation) and at that point it's
probably better to look at full pipelining.

Signed-off-by: Jacob Howard <jacob@mutagen.io>