sync: ignore internal staging directories in watching
Although it's not critical, it's possible for recursive watching
directories to ignore internal staging directories. This didn't happen
automatically because they were still seeing changes within those
directories for files that didn't have the temporary name prefix, but we
can cheaply check both the path and the leaf for this prefix in the
recursive watching case.

Signed-off-by: Jacob Howard <jacob@mutagen.io>