sync: always use volume-internal staging in sidecar
We previously optimized staging in sidecar volumes to be internal if the
synchronization root was a volume root, but we can actually do it for
any descendant of a volume root. This change in implementation actually
allows us to clean up quite a bit of code in the sidecar package and
endpoint initialization.

Signed-off-by: Jacob Howard <jacob@mutagen.io>