sync: always avoid watch re-establishment during Transition
There was already a fortuitous heuristic built in to the recursive
watching loop: we wouldn't attempt to re-establish a recursive watch on
failure before Transition returned because we wait for the scan lock on
watch failure. However, this only occured if acceleration was enabled
(which, fortunately, is almost always the case), so this commit expands
this behavior to include cases where acceleration is disabled.

There's no analagous behavior with poll-based watching, but it's not
nearly as susceptible to watcher overflow due to rapid changes given its
use of (at most) non-recursive watching.

Signed-off-by: Jacob Howard <jacob@mutagen.io>