state: add standalone coalescing implementation
We're going to decompose the local synchronization endpoint
implementation, so the most obvious initial candidate for extraction is
the coalescing logic. We've implemented it here using a background
Goroutine instead of time.AfterFunc because the latter will spawn a
Goroutine every time it fires, which seems wasteful in our context.

Signed-off-by: Jacob Howard <jacob@mutagen.io>