watch: remove in-watcher coalescing and filtering
We originally pushed the coalescing logic out of the endpoint and into
watchers in an effort to simplify the watching and polling run loops in
the endpoint, but, in practice, it ended up making things more
complicated, and we even had to introduce in-watcher filtering to make
it work. It also took two coalescing implementations and turned them
into four implementations.

This commit winds things back a little bit by removing coalescing and
filtering from the watcher implementations, which will re-integrate
into the endpoint with (hopefully) cleaner code (via Coalescer).

This commit also introduces a sentinel error to detect internal watcher
event overflows so that we can make better decisions about watcher
re-establishment in the face of rapid notifications.

Signed-off-by: Jacob Howard <jacob@mutagen.io>