cli: add template-based formatting to list and monitor commands
This commit adds the baseline infrastructure and flags needed for
text/template-based formatting in list and monitor commands. Additional
work is still needed to bring this functionality to other commands and
to provide a more idiomatic (or at least controllable) data context for
templates to use (vs. raw gRPC messages). At the moment, only
text/template built-ins and a custom JSON-encoding built-in are
provided to templates, but this functionality will also be expanded.

Signed-off-by: Jacob Howard <jacob@mutagen.io>