sync: restructure session state and output endpoint statistics
This commit restructures the internal and external session state
representations, encapsulating endpoint-specific state into a separate
structure. This makes it easier to propagate the endpoint content
statistics that are now available in Snapshot, and also cleans up a lot
of the previous code repetition for endpoint-specific state fields.

This commit also modifies synchronization session listings to include
these statistics.

Signed-off-by: Jacob Howard <jacob@mutagen.io>