filesystem: make metadata available via OpenFile methods
Since our (perhaps overly) cautious filesystem traversal code already
pays the price of grabbing filesystem metadata when opening files, it
makes sense to provide this information to the callers of
Directory.OpenFile and Opener.OpenFile. This allows rsync file size
estimation to avoid two seek system calls per file. It comes at the cost
of emulating the standard os package's Windows stat code, but we were
already 90% of the way there.

We haven't included metadata results in Directory.OpenDirectory, because
POSIX systems use the O_DIRECTORY optimization to avoid the cost of an
fstat system call in that case, but if we really wanted to completely
unify the Open* APIs, we could forego that optimization, because it
doesn't seem to make any performance difference (probably because it's
implemented with an fstat call), but at the moment there's no need for
metadata information from Directory.OpenDirectory.

Signed-off-by: Jacob Howard <jacob@mutagen.io>