ssh: add Windows built-in OpenSSH implementation to search paths
This commit adds the official Windows OpenSSH implementation to the SSH
search paths used by Mutagen. This was previously blocked by
PowerShell/Win32-OpenSSH#1152. That issue was resolved in
PowerShell/openssh-portable@675e761, which is available in v8.1.0.0 and
later. It's unclear when exactly v8.1.0.0 actually shipped to Windows
users, but it seems to have been in the latter half of 2021. Since this
is now an officially released version, it makes sense to include it in
Mutagen's SSH search paths.

We're still going to deprioritize this path, because the official
Windows OpenSSH releases tend to be few and far between, not to mention
often delayed by years. As such, it makes sense to continue preferring
the more frequently updated SSH binaries offered by MinGW and Cygwin
distributions.

Signed-off-by: Jacob Howard <jacob@mutagen.io>