all: hide Cobra's built-in completion command
Cobra 1.2.0 added a built-in completion command that slipped under our
radar and is visible in help output. Given that users may have started
using this command, and given that it's the de facto way forward for
Cobra completion, we'll just switch it to be hidden for now.

Eventually, we'll switch over from the V1 Bash completion that our
generate command creates (which seems to work fine with the V2 Bash
completion framework) to a V2-only completion generated by this new
command. However, we'll wait on this for the time being since the
macOS Bash situation is a bit messy.

The issue is that macOS has stopped shipping Bash updates, sticking with
the last GPLv2 version available (3.2). Unfortunately, V2 Bash
completion requires Bash 4.2+. Users can install a more modern Bash via
Homebrew or MacPorts, but that's probably not too common. Since the
current shell completion will work in both scenarios, we'll just keep it
as-is until the macOS Bash situation improves or Apple removes it
entirely. In either case, we'll probably eventually only support V2
(since Cobra is going to deprecate V1 support).

Signed-off-by: Jacob Howard <jacob@mutagen.io>