state: enable polling preemption for state.Tracker
The implementation for state.Tracker previously used only condition
variables, but this didn't jive with the channel-based context
cancellation used with gRPC (and standard HTTP) handlers. To allow for
preemption, this commit adds a background Goroutine to state.Tracker to
serve as an intermediary between channels and condition variables,
allowing for preemptable polling.

Signed-off-by: Jacob Howard <jacob@mutagen.io>