sync: add fully manual permissions propagation mode
While Mutagen's "portable" permissions propagation mode remains the best
default for development, there do exist valid use cases for a fully
manual mode. This commit adds a new PermissionsMode enumeration (along
with the associated machinery) and a new "manual" permissions mode that
allows for fully explicit permission mode setting. This enumeration also
opens the door to future modes such as "posix-raw" and possibly even
"windows-raw". The existing behavior has been designated "portable" for
the time being. The only other sensible name would be "executability",
but that's too tedious and error-prone to type, even if it is the
default.

This commit also fixes a minor bug in transitioner.createSymbolicLink
where the directory mode was being used as the permissions base for
symbolic links. The file mode should have been used instead. The effect
is minor since permission bits on symbolic links aren't used for access
control anyway (except for readlink on macOS, where read permissions are
required to read the symbolic link's target path (though even there the
target itself is still subject to access controls with its own
permission bits)). Also, given that directory permissions are typically
the same as file permissions with executability bits AND'd in, it's
unlikely that any observable differences will manifest in this
particular use case.

Signed-off-by: Jacob Howard <jacob@mutagen.io>