agent: improve command not found detection for some POSIX shells
Some POSIX shells (e.g. tcsh) don't return exit codes 126/127 when a
command isn't found or is invalid. Instead, they return an exit code of
1 and a message such as "command not found". In order to handle these
shells, this commit adds a new detection heuristic for POSIX systems
based on error output.

It doesn't seem to be worth adding this to the Docker transport since it
has such specific error return codes anyway, but it would be an easy
change to make if necessary.

Updates #366

Signed-off-by: Jacob Howard <jacob@mutagen.io>