sync: use file mode as base for symbolic link permissions
This commit fixes a minor bug in transitioner.createSymbolicLink where
the directory mode was being used as the permissions base for symbolic
links. The file mode should have been used instead. The effect is minor
since permission bits on symbolic links aren't used for access control
anyway (except for readlink on macOS, where read permissions are
required to read the symbolic link's target path (though even there the
target itself is still subject to access controls with its own
permission bits)). Also, given that directory permissions are typically
the same as file permissions with executability bits AND'd in, it's
unlikely that any observable differences will manifest in this
particular use case.

This fix is a manual backport of changes in 98c5622.

Signed-off-by: Jacob Howard <jacob@mutagen.io>