sync: fix accelerated scanning corner case for fanotify
The fanotify API never returns paths for content that's created,
deleted, or renamed into place; it only ever returns the parent path of
that content. That's normally not an issue, but there is one case where
that causes problems: when a directory is deleted and another directory
with the same name is renamed into its place. In that case, only the
parent path will be reported. To address this, we can use an additional
heuristic on Linux: we'll consider directories "dirty" (even if they're
not explicitly marked as such) if their immediate parent directory is
marked as dirty.

This won't have any performance impact on accelerated scanning on other
platforms, and the effect on Linux should be minimal. This also has no
impact on any watching, scanning, or polling mechanism beside fanotify.

Files and symbolic links are always rechecked since we pay for their
metadata, so they don't exhibit this issue.

Signed-off-by: Jacob Howard <jacob@mutagen.io>