sync: optimize fanotify edge case heuristic
The edge case addressed by 912bc19 can actually be addressed with almost
zero penalty by using the fact that we know whether or not we're dealing
with an empty baseline directory and the fact that we know the parent
path is marked as dirty. This avoids rescans on almost all directories
that would have been affected as well as an additional map lookup on the
hot path.

Since we've migrated to only using directory baselines, it also made
sense to rename the contentBaseline variable to be clear about that.

Signed-off-by: Jacob Howard <jacob@mutagen.io>