watch: allow internal inotify watcher to drop events
The inotify implementation that we're using doesn't drop events
internally, which can block its internal run loop from signaling
termination if there's a large number of events pending when the watcher
is closed. To solve this, we'll allow the inotify watcher to drop events
since our non-recursive watching is opportunistic/best-effort anyway.

This issue has been manifesting on the Linux/386 builds recently,
causing timeouts due to endpoints not being able to terminate.

Eventually, we'll probably just want to write our own inotify
implementation. Our needs are simpler than those of the code that we've
vendored, which also no longer appears to be maintained upstream.

Signed-off-by: Jacob Howard <jacob@mutagen.io>