fs: use simpler libexec relative path resolution
The os.Executable function already guarantees that it returns an
absolute path (so long as it doesn't return an error), so we can just
perform a filepath.Clean operation instead of filepath.Abs. This has no
effect on the code, other than avoiding an internal filepath.IsAbs check
in filepath.Abs, so it's not worth a backport.

Signed-off-by: Jacob Howard <jacob@mutagen.io>