sync: use 128-bit XXH3 hash for computing path digests in staging
We don't need a cryptographic hash for this purpose, and even though
SHA1 is hardware-accelerated on many platforms, it certainly isn't on
all of them. Thus, it makes sense to switch to a low-collision (but
non-cryptographic) hash that works well everywhere and has the potential
for hardware acceleration. XXH3 fits this bill.

We're also likely going to add support for XXH3 as our content-hashing
function (like rsync has done), in which case adding this dependency is
doubly worth it.

Signed-off-by: Jacob Howard <jacob@mutagen.io>