sync: remove tag-based stager construction
It turns out that the concurrent stager implementation we prototyped on
macOS doesn't provide any significant gain, so there's no point in
providing tag-based controls for switching stager implementations.

That being said, we'll likely continue experimenting with alternative
staging implementations, so we'll leave the reset of the abstractions in
place for now, including the stager interface and separate store type.
If it looks like we won't end up needing this level of abstraction, we
can flatten these in the future.

Signed-off-by: Jacob Howard <jacob@mutagen.io>