all: switch to faster gzip implementation for agent bundle handling
Since we've already integrated the faster FLATE implementation as part
of our staging optimizations, we may as well switch to the corresponding
gzip implementation in both the build script and agent extraction code.

No additional license changes are necessary here since all of this code
is covered by the license text we added when including the compress
package.

Signed-off-by: Jacob Howard <jacob@mutagen.io>