all: disable openbsd/mips64 port
The openbsd/mips64 Go port doesn't seem to have been tested anytime
recently on the Go dashboard and is now causing the Go linker to
segfault with a recent update to the sys subrepository. It's also still
a bit ambiguous as to whether or not it's officially supported. We'll
disable it for now until the picture there clarifies.

Signed-off-by: Jacob Howard <jacob@mutagen.io>