deps: update to Go 1.20
We're going to try something new with this update: we're going to
continue supporting all actively supported releases of Go (i.e. the
current version and the previous version). This shouldn't be too much
additional burden, but it may prove useful as Mutagen's project
landscape continues to expand.

In order to support the previous version, we're going to leave the build
constraint in pkg/mutagen/version.go pointing to the prior release
version and we're going to perform a matrix build on Linux. We could
perform a matrix build on all platforms, but this is merely a canary and
it would be wasteful to tie up more CI builders. In any case, Linux
builds will perform cross-compilation for the most commonly used
platforms, so we should catch any compatibility issues.

Signed-off-by: Jacob Howard <jacob@mutagen.io>