sync: restrict fanotify code inclusion to sidecar
The fanotify code is currently only activated in the sidecar, so it
doesn't make sense to include it in standard Mutagen agent builds. In
theory this might remove the ability for injected agents to take
advantage of this functionality, but in reality the constraints on
fanotify usage are so tight that it only works under very controlled
circumstances, in which case the sidecar is already going to contain an
agent binary matching the Mutagen daemon version. Trying to activate
fanotify functionality on code outside of /volumes would result in
watching failure and a fallback to a very inefficient poll-based
watching, so it's best to avoid that footgun entirely and let injected
agents always default to inotify.

Signed-off-by: Jacob Howard <jacob@mutagen.io>