all: implement Mutagen Pro licensing mechanism for SSPL builds
Most SSPL features (starting with Zstandard and XXH128 support) are
going to require a paid license to use. This license will be the same as
that required to use the Mutagen extension for Docker Desktop. This
change is fundamental to supporting continued Mutagen development. This
commit implements the client-side portion of the licensing mechanism.
This only applies to SSPL builds; no licensing code is included in
MIT-only builds.

Signed-off-by: Jacob Howard <jacob@mutagen.io>