sync: rename hashing algorithm field in YAML and JSON
We're making a slight consistency exception in favor of UX, but I think
it's warranted in this case. In fact, it could be argued that this
nomenclature is still consistent and that the only reason we use more
verbose naming for lesser-used fields it to avoid name collisions.

This commit also adds new YAML test fields to the loading unit test.

Signed-off-by: Jacob Howard <jacob@mutagen.io>