all: don't housekeep agents inside Mutagen sidecar containers
Mutagen sidecar containers are typically unpacked via OverlayFS onto an
ext4 filesystem with either relatime or noatime. In the case of an older
sidecar image, this could result in the agent binary actually unlinking
itself, which won't prevent it from functioning, but which will require
reinstallation by the next session targeting the container (in which
case the fanotify support of the built-in agent will typically be lost).

Signed-off-by: Jacob Howard <jacob@mutagen.io>