cli: neutralize control characters in terminal output
This commit adds neutralization (escaping) of control characters in
terminal output for values that aren't locally controlled. This output
(if unsanitized) could cause terminal corruption. In theory, it could
also be used as an attack vector for intentionally causing terminal
corruption when synchronizing with an untrusted (and malicious) remote.
On very old terminal emulators, it can theoretically be used as an
attack vector for code execution (see, for example, CVE-2003-0069),
though this type of vulnerability has been patched for over 20 years.

We still treat the daemon as a trusted entity (since it is under user
control) and don't sanitize (most) output under its control (such as
locally generated error messages).

We intentionally avoid sanitizing templated output because that output
needs to contain raw, unmodified values. A shellSanitize function has
been added to support neutralization of control characters in template
output.

Signed-off-by: Jacob Howard <jacob@mutagen.io>