sync: adjust approach to ignore pattern validation
The hack that we'd previously put in place for handling permission mode
validation ended up breaking down for ignores. In fact, the approach
wasn't really valid (even for permissions), and we're going to have to
switch to an approach of more general validation until the effective
ignore syntax can be determined.

Signed-off-by: Jacob Howard <jacob@mutagen.io>