sync: use exponential backoff when encountering short-lived loops
When we encounter system-level errors (such as ENOSPC), our best bet is
to keep the run loop alive and continue retrying synchronization.
However, if we try too aggressively, then we end up spinning on the CPU.
Our best bet is to use an exponential backoff strategy driven by the
number of sequential short-lived synchronization loops encountered.

Signed-off-by: Jacob Howard <jacob@mutagen.io>